New structure of Ca, wip
