+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperator.java
@@ -25,7 +25,6 @@ import io.strimzi.api.kafka.model.podset.StrimziPodSet;
 import io.strimzi.certs.CertManager;
 import io.strimzi.operator.cluster.ClusterOperatorConfig;
 import io.strimzi.operator.cluster.PlatformFeaturesAvailability;
-import io.strimzi.operator.cluster.model.ClusterCa;
 import io.strimzi.operator.cluster.model.KafkaCluster;
 import io.strimzi.operator.cluster.model.KafkaVersionChange;
 import io.strimzi.operator.cluster.model.ModelUtils;
@@ -40,12 +39,7 @@ import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationException;
 import io.strimzi.operator.common.ReconciliationLogger;
 import io.strimzi.operator.common.config.ConfigParameter;
-import io.strimzi.operator.common.model.ClientsCa;
-import io.strimzi.operator.common.model.Labels;
-import io.strimzi.operator.common.model.NamespaceAndName;
-import io.strimzi.operator.common.model.PasswordGenerator;
-import io.strimzi.operator.common.model.StatusDiff;
-import io.strimzi.operator.common.model.StatusUtils;
+import io.strimzi.operator.common.model.*;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
@@ -266,8 +260,8 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
 
         /* test */ KafkaVersionChange versionChange;
 
-        /* test */ ClusterCa clusterCa;
-        /* test */ ClientsCa clientsCa;
+        /* test */ Ca clusterCa;
+        /* test */ Ca clientsCa;
 
         // Needed by Cruise control to configure the cluster, its nodes and their storage and resource configuration
         private Set<NodeRef> kafkaBrokerNodes;
