+++ operator-common/src/main/java/io/strimzi/operator/common/model/ClientsCa.java
@@ -5,7 +5,6 @@
 package io.strimzi.operator.common.model;
 
 import io.fabric8.kubernetes.api.model.Secret;
-import io.strimzi.api.kafka.model.common.certmanager.IssuerRef;
 import io.strimzi.certs.CertManager;
 import io.strimzi.operator.common.Reconciliation;
 
@@ -16,7 +15,7 @@ import java.util.Map;
 /**
  * Represents the Strimzi Clients CA
  */
-public class ClientsCa extends Ca {
+public class ClientsCa extends InternalCa {
     /**
      * Creates a ClientsCA instance
      *
@@ -29,28 +28,11 @@ public class ClientsCa extends Ca {
      */
     public ClientsCa(Reconciliation reconciliation, CertManager certManager, PasswordGenerator passwordGenerator,
                      Secret clientsCaCert, Secret clientsCaKey, CaConfig caConfig) {
-        super(reconciliation, certManager, passwordGenerator,
+        super(reconciliation, caConfig, certManager, passwordGenerator,
                 "clients-ca",
                 clientsCaCert,
-                clientsCaKey,
-                caConfig,
-                null);
-    }
-
-    /**
-     * Creates a ClientsCA instance
-     *
-     * @param reconciliation         Reconciliation marker
-     * @param certManager            Certificate manager instance
-     * @param passwordGenerator      Password generator instance
-     * @param clientsCaCert          Kubernetes Secret where the Clients CA public key will be stored
-     * @param clientsCaKey           Kubernetes Secret where the Clients CA private key will be stored
-     * @param caConfig              Certificate Authority configuration
-     * @param issuerRef              Reference to issuer for issuing certificates through other services like cert-manager
-     */
-    public ClientsCa(Reconciliation reconciliation, CertManager certManager, PasswordGenerator passwordGenerator,
-                     Secret clientsCaCert, Secret clientsCaKey, CaConfig caConfig, IssuerRef issuerRef) {
-        super(reconciliation, certManager, passwordGenerator, "clients-ca", clientsCaCert, clientsCaKey, caConfig, issuerRef);
+                clientsCaKey
+        );
     }
 
     @Override
