+++ operator-common/src/main/java/io/strimzi/operator/common/model/InternalCa.java
@@ -1,0 +1,671 @@
+/*
+ * Copyright Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.operator.common.model;
+
+import io.fabric8.kubernetes.api.model.Secret;
+import io.strimzi.certs.CertAndKey;
+import io.strimzi.certs.CertManager;
+import io.strimzi.certs.Subject;
+import io.strimzi.operator.common.Annotations;
+import io.strimzi.operator.common.Reconciliation;
+import io.strimzi.operator.common.ReconciliationLogger;
+import io.strimzi.operator.common.Util;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.security.KeyStoreException;
+import java.security.NoSuchAlgorithmException;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+import java.security.spec.InvalidKeySpecException;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
+import java.util.Base64;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Predicate;
+
+
+/**
+ * A Certificate Authority which can renew its own (self-signed) certificates, and generate signed certificates
+ */
+@SuppressWarnings("checkstyle:CyclomaticComplexity")
+public class InternalCa extends Ca {
+
+    protected static final ReconciliationLogger LOGGER = ReconciliationLogger.create(InternalCa.class);
+
+    /**
+     * Key for storing the CA PKCS21 store in a Kubernetes Secret
+     */
+    public static final String CA_STORE = SecretEntry.P12_KEYSTORE.asKey(CA_SECRET_PREFIX);
+
+    /**
+     * Key for storing the PKCS12 store password in a Kubernetes Secret
+     */
+    public static final String CA_STORE_PASSWORD = SecretEntry.P12_KEYSTORE_PASSWORD.asKey(CA_SECRET_PREFIX);
+
+    /**
+     * Initial generation used for the CAs
+     */
+    public static final int INIT_GENERATION = 0;
+
+    private final PasswordGenerator passwordGenerator;
+    private final CertManager certManager;
+
+    /**
+     * Constructs the CA object
+     *
+     * @param reconciliation    Reconciliation marker
+     * @param caRole
+     * @param certManager       Certificate manager instance
+     * @param passwordGenerator Password generator instance
+     * @param caCertSecret      Kubernetes Secret where the CA public key is stored
+     * @param caKeySecret       Kubernetes Secret where the CA private key is stored
+     * @param caConfig          Certificate Authority configuration
+     */
+    public InternalCa(Reconciliation reconciliation,
+                      CaRole caRole, CertManager certManager,
+                      PasswordGenerator passwordGenerator,
+                      Secret caCertSecret,
+                      Secret caKeySecret,
+                      CaConfig caConfig) {
+        super(reconciliation, caRole, caCertSecret, caKeySecret, caConfig);
+        boolean isGenerateCa = caConfig.isGenerateCa();
+        if (!isGenerateCa && (caCertSecret == null || caKeySecret == null))   {
+            throw new InvalidResourceException(caName() + " should not be generated, but the secrets were not found.");
+        }
+
+        this.certManager = certManager;
+        this.passwordGenerator = passwordGenerator;
+    }
+
+    @Override
+    protected int initCaKeyGeneration(Secret caKeySecret, Secret caCertSecret) {
+        // Strimzi: read from KEY secret
+        if (caKeySecret != null) {
+            return Annotations.intAnnotation(caKeySecret, ANNO_STRIMZI_IO_CA_KEY_GENERATION, INIT_GENERATION);
+        }
+        return INIT_GENERATION;
+    }
+
+    @Override
+    protected Map<String, String> initCaCertData(Secret caCertSecret) {
+        // Strimzi: either generate or validate user-provided
+        if (caConfig.isGenerateCa()) {
+            return caCertSecret == null ? new HashMap<>() : caCertSecret.getData();
+        } else if (caCertSecret != null) {
+            validateUserCaCertChain(caCertSecret.getData());
+            return caCertSecret.getData();
+        }
+        return new HashMap<>();
+    }
+
+    @Override
+    protected Map<String, String> initCaKeyData(Secret caKeySecret) {
+        // Strimzi: copy key data from secret
+        if (caConfig.isGenerateCa()) {
+            return caKeySecret == null ? new HashMap<>() : caKeySecret.getData();
+        } else if (caKeySecret != null) {
+            return Map.of(CA_KEY, caKeySecret.getData().get(CA_KEY));
+        }
+        return new HashMap<>();
+    }
+
+    private static void delete(Reconciliation reconciliation, File file) {
+        if (file != null && !file.delete()) {
+            LOGGER.warnCr(reconciliation, "{} cannot be deleted", file.getName());
+        }
+    }
+
+    /**
+     * Adds a certificate into a PKCS12 keystore
+     *
+     * @param alias     Alias under which it should be stored in the PKCS12 store
+     * @param key       Private key
+     * @param cert      Public key
+     *
+     * @return  PKCS12 store with the certificate
+     *
+     * @throws IOException  Throws an IOException if something fails when working with the files
+     */
+    public CertAndKey addKeyAndCertToKeyStore(String alias, byte[] key, byte[] cert) throws IOException {
+        try {
+            File keyFile = Files.createTempFile("tls", "key").toFile();
+            File certFile = Files.createTempFile("tls", "cert").toFile();
+            File keyStoreFile = null;
+
+            try {
+                Files.write(keyFile.toPath(), key);
+                Files.write(certFile.toPath(), cert);
+
+                if (caConfig.isGeneratePkcs12Stores()) {
+                    keyStoreFile = Files.createTempFile("tls", "p12").toFile();
+
+                    String keyStorePassword = passwordGenerator.generate();
+                    certManager.addKeyAndCertToKeyStore(keyFile, certFile, alias, keyStoreFile, keyStorePassword);
+
+                    return new CertAndKey(
+                            Files.readAllBytes(keyFile.toPath()),
+                            Files.readAllBytes(certFile.toPath()),
+                            null,
+                            Files.readAllBytes(keyStoreFile.toPath()),
+                            keyStorePassword);
+                } else {
+                    return new CertAndKey(
+                            Files.readAllBytes(keyFile.toPath()),
+                            Files.readAllBytes(certFile.toPath()),
+                            null,
+                            null,
+                            null);
+                }
+            } finally {
+                delete(reconciliation, keyFile);
+                delete(reconciliation, certFile);
+                delete(reconciliation, keyStoreFile);
+            }
+        } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException | InvalidKeySpecException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public CertAndKey generateSignedCert(Subject subject,
+                                         File csrFile, File keyFile, File certFile, File keyStoreFile, boolean includeCaChain) {
+        LOGGER.infoCr(reconciliation, "Generating certificate {}, signed by CA {}", subject, this);
+
+        try {
+            byte[] caCertBytes = currentCaCertBytes();
+            certManager.generateCsr(keyFile, csrFile, subject);
+            certManager.generateCert(csrFile, currentCaKey(), caCertBytes,
+                    certFile, subject, caConfig.getValidityDays());
+
+            byte[] certChain;
+            if (includeCaChain) {
+                byte[] leafCert = Files.readAllBytes(certFile.toPath());
+                certChain = new byte[leafCert.length + caCertBytes.length];
+                System.arraycopy(leafCert, 0, certChain, 0, leafCert.length);
+                System.arraycopy(caCertBytes, 0, certChain, leafCert.length, caCertBytes.length);
+            } else {
+                certChain = Files.readAllBytes(certFile.toPath());
+            }
+
+            if (caConfig.isGeneratePkcs12Stores()) {
+                String keyStorePassword = passwordGenerator.generate();
+                certManager.addKeyAndCertToKeyStore(keyFile, certFile, subject.commonName(), keyStoreFile, keyStorePassword);
+
+                return new CertAndKey(
+                        Files.readAllBytes(keyFile.toPath()),
+                        certChain,
+                        null,
+                        Files.readAllBytes(keyStoreFile.toPath()),
+                        keyStorePassword,
+                        caCertGeneration);
+            } else {
+                return new CertAndKey(
+                        Files.readAllBytes(keyFile.toPath()),
+                        certChain,
+                        null,
+                        null,
+                        null,
+                        caCertGeneration);
+            }
+        } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException | InvalidKeySpecException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Generates a certificate signed by this CA
+     *
+     * @param commonName The CN of the certificate to be generated.
+     * @return A CompletionStage which completes with the newly generated certificate CertAndKey
+     * @throws IOException If the cert could not be generated.
+     */
+    public CertAndKey generateSignedCert(String commonName) throws IOException {
+        return generateSignedCert(getSubject(commonName, null));
+    }
+
+    private Subject getSubject(String commonName, String organization) {
+        Subject.Builder subject = new Subject.Builder();
+        if (organization != null) {
+            subject.withOrganizationName(organization);
+        }
+        subject.withCommonName(commonName);
+        return subject.build();
+    }
+
+    /**
+     * Gets a certificate signed by this CA
+     *
+     * @param commonName The CN of the certificate to be created.
+     * @param organization The O of the certificate to be created. May be null.
+     * @return The newly created certificate CertAndKey
+     * @throws IOException If the cert could not be created.
+     */
+    public CertAndKey getSignedCert(String commonName, String organization) throws IOException {
+        Subject subject = getSubject(commonName, organization);
+        return generateSignedCert(subject);
+    }
+
+    /**
+     * Generates a certificate signed by this CA
+     *
+     * @param subject The subject of the certificate to be generated.
+     * @return The CertAndKey
+     * @throws IOException If the cert could not be generated.
+     */
+    private CertAndKey generateSignedCert(Subject subject) throws IOException {
+        File csrFile = Files.createTempFile("tls", "csr").toFile();
+        File keyFile = Files.createTempFile("tls", "key").toFile();
+        File certFile = Files.createTempFile("tls", "cert").toFile();
+        File keyStoreFile = Files.createTempFile("tls", "p12").toFile();
+
+        CertAndKey result = generateSignedCert(subject, csrFile, keyFile, certFile, keyStoreFile, false);
+
+        delete(reconciliation, csrFile);
+        delete(reconciliation, keyFile);
+        delete(reconciliation, certFile);
+        delete(reconciliation, keyStoreFile);
+        return result;
+    }
+
+    /**
+     * Returns whether the certificate is expiring or not
+     *
+     * @param secret  Secret with the certificate
+     * @param certKey   Key under which is the certificate stored
+     * @return  True when the certificate should be renewed. False otherwise.
+     */
+    public boolean isExpiring(Secret secret, String certKey)  {
+        X509Certificate currentCert = CaUtils.cert(secret, certKey);
+        return certNeedsRenewal(currentCert);
+    }
+
+    public boolean isGenerateCa()  {
+        return caConfig.isGenerateCa();
+    }
+
+
+    /**
+     * Returns whether the certificate is expiring or not
+     *
+     * @param certificate Byte array with the certificate
+     *
+     * @return  True when the certificate should be renewed. False otherwise.
+     */
+    public boolean isExpiring(byte[] certificate)  {
+        try {
+            X509Certificate currentCert = CaUtils.x509Certificate(certificate);
+            return certNeedsRenewal(currentCert);
+        } catch (CertificateException e) {
+            LOGGER.errorCr(reconciliation, "Failed to parse existing certificate", e);
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Create or update CA data when Strimzi is managing CA.
+     * <p>
+     * Create the CA data if they don't exist, otherwise if within the renewal period then either renew
+     * the CA cert or replace the CA cert and key, according to the configured policy. After calling this method
+     * {@link #certsRemoved()} will return whether expired secrets were removed from the Secret.
+     *
+     * @param maintenanceWindowSatisfied Flag indicating whether we are in the maintenance window
+     * @param forceReplace Flag indicating whether to do a force replace
+     * @param forceRenew Flag indicating whether to do a force renew
+     */
+    public void createOrUpdateStrimziManagedCa(boolean maintenanceWindowSatisfied, boolean forceReplace, boolean forceRenew) {
+        X509Certificate currentCert = currentCaCertX509();
+        Map<String, String> certData;
+        Map<String, String> keyData;
+        this.renewalType = shouldCreateOrRenewStrimziManagedCa(currentCert, maintenanceWindowSatisfied, forceReplace, forceRenew);
+        LOGGER.debugCr(reconciliation, "{} renewalType {}", this, renewalType);
+
+        switch (renewalType) {
+            case CREATE -> {
+                keyData = new HashMap<>(1);
+                certData = new HashMap<>(3);
+                generateCaKeyAndCert(nextCaSubject(caKeyGeneration), keyData, certData);
+            }
+            case REPLACE_KEY -> {
+                keyData = new HashMap<>(1);
+                certData = new HashMap<>(caCertData);
+                if (certData.containsKey(CA_CRT)) {
+                    String notAfterDate = DATE_TIME_FORMATTER.format(currentCert.getNotAfter().toInstant().atZone(ZoneId.of("Z")));
+
+                    if (caConfig.isGeneratePkcs12Stores()) {
+                        addCertCaToTrustStore("ca-" + notAfterDate + SecretEntry.CRT.suffix, certData);
+                    }
+
+                    certData.put("ca-" + notAfterDate + SecretEntry.CRT.suffix, certData.remove(CA_CRT));
+                }
+                ++caCertGeneration;
+                generateCaKeyAndCert(nextCaSubject(++caKeyGeneration), keyData, certData);
+            }
+            case RENEW_CERT -> {
+                keyData = new HashMap<>(caKeyData);
+                certData = new HashMap<>(3);
+                ++caCertGeneration;
+                renewCaCert(nextCaSubject(caKeyGeneration), certData);
+            }
+            default -> {
+                keyData = new HashMap<>(caKeyData);
+                certData = new HashMap<>(caCertData);
+
+                if (caConfig.isGeneratePkcs12Stores() && !certData.containsKey(CA_STORE)) {
+                    // If we are generating PKCS12 stores, and it is missing in the Secret, we add it
+                    addCertCaToTrustStore(CA_CRT, certData);
+                } else if (!caConfig.isGeneratePkcs12Stores() && certData.containsKey(CA_STORE)) {
+                    // If we are not generating PKCS12 stores, and it is already present in the Secret, we will remove it
+                    certData.remove(CA_STORE);
+                    certData.remove(CA_STORE_PASSWORD);
+                }
+            }
+        }
+
+        if (removeCerts(certData, this::removeExpiredCert)) {
+            LOGGER.infoCr(reconciliation, "{}: Expired CA certificates removed", this);
+            this.caCertsRemoved = true;
+        }
+
+        if (renewalType != RenewalType.NOOP && renewalType != RenewalType.POSTPONED) {
+            LOGGER.debugCr(reconciliation, "{}: {}", this, renewalType.postDescription(caName())) ;
+        }
+        caCertData = certData;
+        caKeyData = keyData;
+    }
+
+    private Subject nextCaSubject(int version) {
+        return new Subject.Builder()
+        // Key replacements does not work if both old and new CA certs have the same subject DN, so include the
+        // key generation in the DN so the certificates appear distinct during CA key replacement.
+            .withCommonName(caName() + " v" + version)
+            .withOrganizationName(IO_STRIMZI).build();
+    }
+
+    private RenewalType shouldCreateOrRenewStrimziManagedCa(X509Certificate currentCert, boolean maintenanceWindowSatisfied, boolean forceReplace, boolean forceRenew) {
+        String reason = null;
+        RenewalType renewalType = RenewalType.NOOP;
+        if (caKeyData.get(CA_KEY) == null) {
+            reason = "CA key secret for " + caName()  + " is missing or lacking data." + CA_KEY.replace(".", "\\.");
+            renewalType = RenewalType.CREATE;
+        } else if (this.caCertData.get(CA_CRT) == null) {
+            reason = "CA certificate secret for " + caName()  + " is missing or lacking data." + CA_CRT.replace(".", "\\.");
+            renewalType = RenewalType.RENEW_CERT;
+        } else if (forceRenew) {
+            reason = "CA certificate secret for " + caName()  + " is annotated with " + Annotations.ANNO_STRIMZI_IO_FORCE_RENEW;
+
+            if (maintenanceWindowSatisfied) {
+                renewalType = RenewalType.RENEW_CERT;
+            } else {
+                renewalType = RenewalType.POSTPONED;
+            }
+        } else if (forceReplace) {
+            reason = "CA key secret for " + caName()  + " is annotated with " + Annotations.ANNO_STRIMZI_IO_FORCE_REPLACE;
+
+            if (maintenanceWindowSatisfied) {
+                renewalType = RenewalType.REPLACE_KEY;
+            } else {
+                renewalType = RenewalType.POSTPONED;
+            }
+        } else if (currentCert != null
+                && certNeedsRenewal(currentCert)) {
+            reason = "Within renewal period for CA certificate (expires on " + currentCert.getNotAfter() + ")";
+
+            if (maintenanceWindowSatisfied) {
+                renewalType = switch (caConfig.getCertificateExpirationPolicy()) {
+                    case REPLACE_KEY -> RenewalType.REPLACE_KEY;
+                    case RENEW_CERTIFICATE -> RenewalType.RENEW_CERT;
+                };
+            } else {
+                renewalType = RenewalType.POSTPONED;
+            }
+        }
+
+        switch (renewalType) {
+            case NOOP ->
+                    LOGGER.debugCr(reconciliation, "{}: {}", this, renewalType.preDescription(caName())) ;
+            case REPLACE_KEY, RENEW_CERT, CREATE ->
+                    LOGGER.debugCr(reconciliation, "{}: {}: {}", this, renewalType.preDescription(caName()) , reason);
+            case POSTPONED ->
+                    LOGGER.warnCr(reconciliation, "{}: {}: {}", this, renewalType.preDescription(caName()) , reason);
+        }
+
+        return renewalType;
+    }
+
+    /**
+     * Remove old certificates that are stored in the CA Secret matching the "ca-YYYY-MM-DDTHH-MM-SSZ.crt" naming pattern.
+     * NOTE: mostly used when a CA certificate is renewed by replacing the key
+     */
+    @Override
+    public void maybeDeleteOldCerts() {
+        // the operator doesn't have to touch Secret provided by the user with his own custom CA certificate
+        if (isGenerateCa()) {
+            deleteOldCerts();
+        }
+    }
+
+    /**
+     * Checks if the renewal or replacement was postponed.
+     *
+     * @return Returns true if the renewal or replacement was postponed
+     */
+    public boolean postponed() {
+        return renewalType.equals(RenewalType.POSTPONED);
+    }
+
+    /**
+     * Predicate used to remove expired certificates that are stored in the CA Secret
+     *
+     * @param entry entry in the CA Secret data section to check
+     * @return if the certificate is expired and has to be removed
+     */
+    private boolean removeExpiredCert(Map.Entry<String, String> entry) {
+        boolean remove = false;
+        String certName = entry.getKey();
+        String certText = entry.getValue();
+        try {
+            X509Certificate cert = CaUtils.x509Certificate(Util.decodeBytesFromBase64(certText));
+            Instant expiryDate = cert.getNotAfter().toInstant();
+            remove = expiryDate.isBefore(clock.instant());
+            if (remove) {
+                LOGGER.infoCr(reconciliation, "The certificate (data.{}) in Secret expired {}; removing it",
+                        certName.replace(".", "\\."), expiryDate);
+            }
+        } catch (CertificateException e) {
+            // doesn't remove stores and related password
+            if (!SecretEntry.P12_KEYSTORE.matchesType(certName) && !SecretEntry.P12_KEYSTORE_PASSWORD.matchesType(certName)) {
+                remove = true;
+                LOGGER.debugCr(reconciliation, "The certificate (data.{}) in Secret is not an X.509 certificate; removing it",
+                        certName.replace(".", "\\."));
+            }
+        }
+        return remove;
+    }
+
+    @Override
+    public boolean removeCerts(Map<String, String> newData, Predicate<Map.Entry<String, String>> predicate) {
+        Iterator<Map.Entry<String, String>> iter = newData.entrySet().iterator();
+        List<String> removed = new ArrayList<>();
+        while (iter.hasNext()) {
+            Map.Entry<String, String> entry = iter.next();
+            boolean remove = predicate.test(entry);
+            if (remove) {
+                String certName = entry.getKey();
+                LOGGER.debugCr(reconciliation, "Removing data.{} from Secret",
+                        certName.replace(".", "\\."));
+                iter.remove();
+                removed.add(certName);
+            }
+        }
+        if (removed.isEmpty()) {
+            return false;
+        } else {
+            if (caConfig.isGeneratePkcs12Stores()) {
+                // the certificates removed from the Secret data have to be removed from the store as well
+                try {
+                    File trustStoreFile = Files.createTempFile("tls", "-truststore").toFile();
+                    Files.write(trustStoreFile.toPath(), Util.decodeBytesFromBase64(newData.get(CA_STORE)));
+                    try {
+                        String trustStorePassword = Util.decodeFromBase64(newData.get(CA_STORE_PASSWORD));
+                        certManager.deleteFromTrustStore(removed, trustStoreFile, trustStorePassword);
+                        newData.put(CA_STORE, Base64.getEncoder().encodeToString(Files.readAllBytes(trustStoreFile.toPath())));
+                    } finally {
+                        delete(reconciliation, trustStoreFile);
+                    }
+                } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException e) {
+                    throw new RuntimeException(e);
+                }
+            }
+
+            return true;
+        }
+    }
+
+    private boolean certNeedsRenewal(X509Certificate cert)  {
+        Instant notAfter = cert.getNotAfter().toInstant();
+        Instant renewalPeriodBegin = notAfter.minus(caConfig.getRenewalDays(), ChronoUnit.DAYS);
+        LOGGER.traceCr(reconciliation, "Certificate {} expires on {} renewal period begins on {}", cert.getSubjectX500Principal(), notAfter, renewalPeriodBegin);
+        return this.clock.instant().isAfter(renewalPeriodBegin);
+    }
+
+
+    private void addCertCaToTrustStore(String alias, Map<String, String> certData) {
+        try {
+            File certFile = Files.createTempFile("tls", "-cert").toFile();
+            Files.write(certFile.toPath(), Util.decodeBytesFromBase64(certData.get(CA_CRT)));
+            try {
+                File trustStoreFile = Files.createTempFile("tls", "-truststore").toFile();
+                if (certData.containsKey(CA_STORE)) {
+                    Files.write(trustStoreFile.toPath(), Util.decodeBytesFromBase64(certData.get(CA_STORE)));
+                }
+                try {
+                    String trustStorePassword = certData.containsKey(CA_STORE_PASSWORD) ?
+                            Util.decodeFromBase64(certData.get(CA_STORE_PASSWORD)) :
+                            passwordGenerator.generate();
+                    certManager.addCertToTrustStore(certFile, alias, trustStoreFile, trustStorePassword);
+                    certData.put(CA_STORE, Base64.getEncoder().encodeToString(Files.readAllBytes(trustStoreFile.toPath())));
+                    certData.put(CA_STORE_PASSWORD, Base64.getEncoder().encodeToString(trustStorePassword.getBytes(StandardCharsets.US_ASCII)));
+                } finally {
+                    delete(reconciliation, trustStoreFile);
+                }
+            } finally {
+                delete(reconciliation, certFile);
+            }
+
+        } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void generateCaKeyAndCert(Subject subject, Map<String, String> keyData, Map<String, String> certData) {
+        try {
+            LOGGER.infoCr(reconciliation, "Generating CA with subject={}", subject);
+
+            File keyFile = Files.createTempFile("tls", subject.commonName() + "-key").toFile();
+            File certFile = Files.createTempFile("tls", subject.commonName() + "-cert").toFile();
+
+            try {
+                certManager.generateSelfSignedCert(keyFile, certFile, subject, caConfig.getValidityDays());
+                CertAndKey ca;
+
+                if (caConfig.isGeneratePkcs12Stores()) {
+                    File trustStoreFile = Files.createTempFile("tls", subject.commonName() + "-truststore").toFile();
+                    String trustStorePassword;
+                    // if secret already contains the truststore, we have to reuse it without changing password
+                    if (certData.containsKey(CA_STORE)) {
+                        Files.write(trustStoreFile.toPath(), Util.decodeBytesFromBase64(certData.get(CA_STORE)));
+                        trustStorePassword = Util.decodeFromBase64(certData.get(CA_STORE_PASSWORD));
+                    } else {
+                        trustStorePassword = passwordGenerator.generate();
+                    }
+                    try {
+                        certManager.addCertToTrustStore(certFile, CA_CRT, trustStoreFile, trustStorePassword);
+                        ca = new CertAndKey(
+                                Files.readAllBytes(keyFile.toPath()),
+                                Files.readAllBytes(certFile.toPath()),
+                                Files.readAllBytes(trustStoreFile.toPath()),
+                                null,
+                                trustStorePassword);
+
+                        certData.put(CA_STORE, ca.trustStoreAsBase64String());
+                        certData.put(CA_STORE_PASSWORD, ca.storePasswordAsBase64String());
+                    } finally {
+                        delete(reconciliation, trustStoreFile);
+                    }
+                } else {
+                    ca = new CertAndKey(
+                            Files.readAllBytes(keyFile.toPath()),
+                            Files.readAllBytes(certFile.toPath()),
+                            null,
+                            null,
+                            null);
+                }
+
+                certData.put(CA_CRT, ca.certAsBase64String());
+                keyData.put(CA_KEY, ca.keyAsBase64String());
+            } finally {
+                delete(reconciliation, certFile);
+                delete(reconciliation, keyFile);
+            }
+        } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void renewCaCert(Subject subject, Map<String, String> certData) {
+        try {
+            LOGGER.infoCr(reconciliation, "Renewing CA with subject={}", subject);
+
+            byte[] bytes = Util.decodeBytesFromBase64(caKeyData.get(CA_KEY));
+            File keyFile = Files.createTempFile("tls", subject.commonName() + "-key").toFile();
+            Files.write(keyFile.toPath(), bytes);
+            File certFile = Files.createTempFile("tls", subject.commonName() + "-cert").toFile();
+
+            try {
+                certManager.renewSelfSignedCert(keyFile, certFile, subject, caConfig.getValidityDays());
+                CertAndKey ca;
+
+                if (caConfig.isGeneratePkcs12Stores()) {
+                    File trustStoreFile = Files.createTempFile("tls", subject.commonName() + "-truststore").toFile();
+                    try {
+                        String trustStorePassword = passwordGenerator.generate();
+                        certManager.addCertToTrustStore(certFile, CA_CRT, trustStoreFile, trustStorePassword);
+                        ca = new CertAndKey(
+                                bytes,
+                                Files.readAllBytes(certFile.toPath()),
+                                Files.readAllBytes(trustStoreFile.toPath()),
+                                null,
+                                trustStorePassword);
+                        certData.put(CA_STORE, ca.trustStoreAsBase64String());
+                        certData.put(CA_STORE_PASSWORD, ca.storePasswordAsBase64String());
+                    } finally {
+                        delete(reconciliation, trustStoreFile);
+                    }
+                } else {
+                    ca = new CertAndKey(
+                            bytes,
+                            Files.readAllBytes(certFile.toPath()),
+                            null,
+                            null,
+                            null);
+                }
+
+                certData.put(CA_CRT, ca.certAsBase64String());
+            } finally {
+                delete(reconciliation, keyFile);
+                delete(reconciliation, certFile);
+            }
+        } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
