+++ user-operator/src/main/java/io/strimzi/operator/user/model/KafkaUserModel.java
@@ -25,7 +25,7 @@ import io.strimzi.operator.common.Annotations;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
 import io.strimzi.operator.common.Util;
-import io.strimzi.operator.common.model.Ca;
+import io.strimzi.operator.common.model.InternalCa;
 import io.strimzi.operator.common.model.CaConfig;
 import io.strimzi.operator.common.model.ClientsCa;
 import io.strimzi.operator.common.model.InvalidResourceException;
@@ -248,7 +248,7 @@ public class KafkaUserModel {
                 passwordGenerator,
                 clientsCaCertSecret,
                 clientsCaKeySecret,
-                new CaConfig(validityDays, renewalDays, false, generatePkcs12Stores, CertificateManagerType.STRIMZI_IO)
+                new CaConfig(validityDays, renewalDays, false, generatePkcs12Stores, CertificateManagerType.STRIMZI_IO, )
         );
         this.caCert = clientsCa.currentCaCertBase64();
 
@@ -295,16 +295,16 @@ public class KafkaUserModel {
         }
     }
 
-    CertAndKey generateNewCertificate(Reconciliation reconciliation, Ca clientsCa) {
+    CertAndKey generateNewCertificate(Reconciliation reconciliation, InternalCa clientsStrimziCa) {
         try {
-            return clientsCa.generateSignedCert(name);
+            return clientsStrimziCa.generateSignedCert(name);
         } catch (IOException e) {
             LOGGER.errorCr(reconciliation, "Error generating signed certificate for user {}", name, e);
             return null;
         }
     }
 
-    CertAndKey reuseCertificate(Reconciliation reconciliation, Ca clientsCa, Secret userSecret) {
+    CertAndKey reuseCertificate(Reconciliation reconciliation, InternalCa clientsStrimziCa, Secret userSecret) {
         String userKeyStore = userSecret.getData().get("user.p12");
         String userKeyStorePassword = userSecret.getData().get("user.password");
 
@@ -321,7 +321,7 @@ public class KafkaUserModel {
         } else {
             // coming from an older operator version, the user secret exists but without keystore and password
             try {
-                return clientsCa.addKeyAndCertToKeyStore(name,
+                return clientsStrimziCa.addKeyAndCertToKeyStore(name,
                         decodeFromSecret(userSecret, "user.key"),
                         decodeFromSecret(userSecret, "user.crt"));
             } catch (IOException e) {
